refactor: improvements on addr package

+++ addr/port.go
@@ -2,10 +2,10 @@ package addr
 
 import "strconv"
 
-func ParsePort(port string) (uint16, error) {
+func ParsePort(port string) (int, error) {
 	portNum, err := strconv.ParseUint(port, 10, 16)
 	if err != nil {
 		return 0, err
 	}
-	return uint16(portNum), nil
+	return int(portNum), nil
 }
